Finish StoreRepositoryDaoImp implementation

+++ src/main/java/com/example/goodstrack/repositories/SupplierRepository.java
@@ -4,6 +4,6 @@ import com.example.goodstrack.domain.Supplier;
 import java.util.Optional;
 
 public interface SupplierRepository {
-    Optional<Supplier> findById(Long id);
+    Optional<Supplier> findById(int id);
     Supplier save(Supplier supplier);
 }
